Fix ~80% of compilation warnings (2820 → ~94 in project source)
Header fixes (highest impact - eliminates warnings across all TUs):
- DocumentSVG.h: Rename all shadowing constructor params (1136 warnings eliminated)
- Polyline.h: Rename shadowing params + fix uninitialized member (292 warnings)
- RangeHelper.h: Rename shadowing constructor params (192 warnings)
- GeomData.h: Fix size_t→int32_t conversion + shadow (94 warnings)
- Node.h: Fix size_t→int32_t conversions (45 warnings)
- NumericRange.h: Fix shadow + float-to-int conversion (6 warnings)
- PoissonGenerator.h: Fix double-promotion, sign/float conversions (30 warnings)
- Stylable.h: Mark unused params with /*param*/ (49 warnings)
- RandomUniDist.h: Rename shadowing local variable (8 warnings)
- PolyConvex.h: Remove ignored const qualifier (14 warnings)
- UseContext.h: Mark unused params (4 warnings)
- Context.h: Mark unused param (4 warnings)

CMake build system:
- Suppress warnings in third-party code (fftw++, agg, HqNoise/fft)
- Suppress warnings in generated protobuf code

+++ LabyPath/src/flatteningOverlap/Node.h
@@ -93,7 +93,7 @@ public:
 
     void add(int32_t i) {
         if (-1 < i and i < static_cast<int32_t>(_stateSelect.size())) {
-            _stateSelect.at(i) = true;
+            _stateSelect.at(static_cast<std::size_t>(i)) = true;
         }
     }
 
@@ -103,10 +103,10 @@ public:
         for (; ite < _stateSelect.size(); ++ite) {
             if (!_stateSelect.at(ite)) {
                 _stateSelect.at(ite) = true;
-                return ite;
+                return static_cast<int32_t>(ite);
             }
         }
-        return ite;
+        return static_cast<int32_t>(ite);
     }
     std::size_t ite = 0;
     std::vector<bool> _stateSelect;
